feat(desktop): show dropzone on new sessions
